+++ src/test/unit/key-generator.test.js
@@ -12,7 +12,7 @@ let Morfix,
   invalidContext,
   KeyGeneratorObject = null
 beforeAll(async () => {
-  Morfix = await Seal
+  Morfix = await Seal()
   const lib = getLibrary()
   KeyGeneratorObject = KeyGenerator(lib)(Morfix)