+++ CHANGES.md
@@ -3,6 +3,19 @@
 See [Microsoft's Change log](https://github.com/microsoft/SEAL/blob/master/Changes.md)
 for more details on each SEAL version change.
 
+## Version 4.2.0
+
+Breaking:
+- Minor breaking change to allow for functional closure. 
+
+```javascript
+import { Seal } from 'node-seal'
+...
+const morfix = await Seal // Deprecated
+const morfix = await Seal() // Use new closure constructor
+```
+
+
 ## Version 4.1.5
 
 Feat: